requirements.txt
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.1
Automat==22.10.0
cffi==1.15.1
channels==3.0.4
constantly==15.1.0
cryptography==41.0.1
daphne==3.0.2
Django==4.1.9
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pymongo==3.12.1
pyOpenSSL==23.2.0
pytz==2023.3
service-identity==21.1.0
six==1.16.0
sqlparse==0.2.4
Twisted==22.10.0
twisted-iocpsupport==1.0.3
txaio==23.1.1
typing_extensions==4.6.3
tzdata==2023.3
zope.interface==6.0